Move sidebar to its own file
